# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "InvitationCreated": {
        "properties": {
          "accept_path": {
            "description": "Where the token goes, so a client does not have to build the URL itself\nand get it subtly wrong.\n\nOn the **apex**, not the tenant's subdomain: the recipient may have no\naccount anywhere yet.",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "handle": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "role": {
            "description": "One of `owner`, `accountant`, `clerk`, `viewer`.",
            "type": "string"
          },
          "token": {
            "description": "**Shown once.** Pass it on however you already talk to this person.\nNothing stores it, so a lost link is re-issued rather than recovered.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "handle",
          "role",
          "expires_at",
          "token",
          "accept_path"
        ],
        "type": "object"
      },
      "InvitationView": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "handle": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "role": {
            "description": "One of `owner`, `accountant`, `clerk`, `viewer`.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "handle",
          "role",
          "expires_at",
          "created_at"
        ],
        "type": "object"
      },
...
      "NewInvitation": {
        "example": {
          "handle": "khalid@acme.example",
          "role": "accountant"
        },
        "properties": {
          "handle": {
            "description": "The address to invite. Also the login handle they will end up with.",
            "type": "string"
          },
          "role": {
            "description": "One of `owner`, `accountant`, `clerk`, `viewer`.",
            "type": "string"
          }
        },
        "required": [
          "handle",
          "role"
        ],
        "type": "object"
      },
...
      "Problem": {
        "example": {
          "code": "auth.invalid_credentials",
          "detail": "Those login details are not right. Please try again.",
          "status": 401,
          "title": "Unauthorized",
          "type": "https://errors.spa.example/auth.invalid_credentials"
        },
        "properties": {
          "args": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MessageArg"
            },
            "description": "The values the message names, typed — which account, which module, how\nmuch was outstanding. A client that renders its own sentence reads these\ninstead of parsing `detail`. Absent when the message takes none.",
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "code": {
            "description": "The stable identifier. Branch on this, never on `detail`.",
            "type": "string"
          },
          "detail": {
            "description": "The message in the caller's language, chosen by `Accept-Language`.",
            "type": "string"
          },
          "status": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "description": "`https://errors.spa.example/` followed by the code. Stable, and a place\nto look it up — see `docs/ERRORS.md` for every code this API can answer\nwith.",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session": {
        "description": "A token from `POST /v1/sessions`. Sessions expire; a 401 with `auth.session_expired` means log in again.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "license": {
      "name": "AGPL-3.0-or-later"
    },
    "title": "SPA",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/invitations": {
      "get": {
        "operationId": "list_invitations",
        "parameters": [
          {
            "description": "The tenant's subdomain — `bassat.spa.com`. Every path below is about that tenant.",
            "in": "header",
            "name": "Host",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvitationView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Pending invitations. The tokens are not here — they were shown once."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No session, or one that has expired. Log in again."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Your role here does not permit this."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such tenant, or not one of yours — the same answer for both, so this API is not a tenant-enumeration oracle."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Something went wrong here. `code` is `system.internal_error`, and nothing more specific is safe to infer."
          }
        },
        "summary": "Invitations sent and not yet taken up.",
        "tags": [
          "invitations"
        ]
      },
      "parameters": [
        {
          "description": "`ar` or `en`. Anything else falls back to English. Honoured on errors too.",
          "in": "header",
          "name": "Accept-Language",
          "required": false,
          "schema": {
            "examples": [
              "ar"
            ],
            "type": "string"
          }
        }
      ],
      "post": {
        "description": "The `token` comes back **once** and is stored nowhere — pass it on however\nyou already talk to this person. A lost link is re-issued, not recovered.",
        "operationId": "invite",
        "parameters": [
          {
            "description": "The tenant's subdomain — `bassat.spa.com`. Every path below is about that tenant.",
            "in": "header",
            "name": "Host",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewInvitation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationCreated"
                }
              }
            },
            "description": "Created."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such role"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No session, or one that has expired. Log in again."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Your role here does not permit this."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such tenant, or not one of yours — the same answer for both, so this API is not a tenant-enumeration oracle."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Already a member here"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "The body was sent without `Content-Type: application/json`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "The body is JSON and not the shape this route takes. `args.reason` names the field."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Something went wrong here. `code` is `system.internal_error`, and nothing more specific is safe to infer."
          }
        },
        "summary": "Invite somebody, leaving the password to them.",
        "tags": [
          "invitations"
        ]
      }
    },
    "/v1/invitations/{invitation}": {
      "delete": {
        "operationId": "revoke_invitation",
        "parameters": [
          {
            "description": "The tenant's subdomain — `bassat.spa.com`. Every path below is about that tenant.",
            "in": "header",
            "name": "Host",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "From `GET /v1/invitations`.",
            "in": "path",
            "name": "invitation",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Withdrawn. The token no longer works."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not a well-formed id"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No session, or one that has expired. Log in again."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Your role here does not permit this."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such tenant, or not one of yours — the same answer for both, so this API is not a tenant-enumeration oracle."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Something went wrong here. `code` is `system.internal_error`, and nothing more specific is safe to infer."
          }
        },
        "summary": "Withdraw an invitation before it is taken up.",
        "tags": [
          "invitations"
        ]
      },
      "parameters": [
        {
          "description": "`ar` or `en`. Anything else falls back to English. Honoured on errors too.",
          "in": "header",
          "name": "Accept-Language",
          "required": false,
          "schema": {
            "examples": [
              "ar"
            ],
            "type": "string"
          }
        }
      ]
    },
...
    "/v1/ledger/chart": {
      "parameters": [
        {
          "description": "`ar` or `en`. Anything else falls back to English. Honoured on errors too.",
          "in": "header",
          "name": "Accept-Language",
          "required": false,
          "schema": {
            "examples": [
              "ar"
            ],
            "type": "string"
          }
        }
      ],
      "post": {
        "description": "Installing twice is not an error: accounts that are already open are counted\nas `skipped` and left exactly as they are, names included.",
        "operationId": "install_chart",
        "parameters": [
          {
            "description": "The tenant's subdomain — `bassat.spa.com`. Every path below is about that tenant.",
            "in": "header",
            "name": "Host",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallChart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartInstalled"
                }
              }
            },
            "description": "Done."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such chart, or an unknown currency"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No session, or one that has expired. Log in again."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Your role here does not permit this."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such tenant, or not one of yours — the same answer for both, so this API is not a tenant-enumeration oracle."
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "The body was sent without `Content-Type: application/json`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "The body is JSON and not the shape this route takes. `args.reason` names the field."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Something went wrong here. `code` is `system.internal_error`, and nothing more specific is safe to infer."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Backpressure. Retryable."
          }
        },
        "summary": "Open every account in a ready-made chart.",
        "tags": [
          "ledger"
        ]
      }
    },
...
  },
  "security": [
    {
      "session": []
    }
  ],
...
}